requirements.txt
langchain
PyPDF2
openai
wave
streamlit==1.25.0
tiktoken
nltk
#azure-cognitiveservices-speech
audio_recorder_streamlit
streamlit-option-menu
streamlit-lottie
faiss-cpu
boto3
Ipython
python-dotenv